pyproject.toml
[project]
name = "byteart"
version = "0.1.0"
description = "encode bytes to randomized pixel arts"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "pytest>=8.4.0",
]

[tool.uv]
native-tls = true
